+++ src/components/PostCard.tsx
@@ -16,7 +16,7 @@ export default function PostCard({ post }: { post: PostProps }) {
       <section className="py-2 px-3 text-center">
         <p className="text-right text-[0.6em] sm:text-xs text-secondary font-light">{date}</p>
         <div className="flex flex-col items-center">
-          <h2 className="text-[0.8em] sm:text-sm font-medium py-1 truncate">{title}</h2>
+          <h2 className="text-[0.8em] sm:text-sm font-medium py-1 w-full truncate">{title}</h2>
           <p className="text-[0.6em] sm:text-xs font-light truncate w-full">{description}</p>
           <span className="text-[0.6em] sm:text-xs rounded-lg bg-point px-2 py-0.5 my-3">{category}</span>
         </div>
